+++ mitra_api/src/mastodon_api/lists/views.rs
@@ -195,8 +195,9 @@ async fn add_accounts_to_list(
     auth: BearerAuth,
     db_pool: web::Data<DatabaseConnectionPool>,
     list_id: web::Path<i32>,
-    accounts_data: web::Json<ListAccountsData>,
+    accounts_data: JsonOrForm<ListAccountsData>,
 ) -> Result<HttpResponse, MastodonError> {
+    let accounts_data = accounts_data.into_inner();
     let db_client = &**get_database_client(&db_pool).await?;
     let current_user = get_current_user(db_client, auth.token()).await?;
     let feed = get_custom_feed(