+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Support POSTing form data to `/api/v1/lists/{list_id}/accounts` endpoint.
+
 ### Changed
 
 - Disabled early filtering for incoming activities with ap:// IDs.